+++ docs/release-notes.md
@@ -17,9 +17,10 @@
 
 ## 2.5.4
 * For players:
-  * Fixed NPC and tree textures not updated when changed through the content API.
+  * Fixed fence, NPC, and tree textures not updated when a mod changes them through the content API.
   * Fixed visual bug on Linux/Mac when mods overlay images through the content API.
   * Fixed error when a mod removes an asset editor/loader.
+  * Fixed minimum game version incorrectly changed from 1.2.30 to 1.2.33 in SMAPI 2.5.3.
 
 * For the [log parser][]:
   * Fixed error when log text contains certain tokens.